optimize: - add serializers for articles, comments - add common response serializer - optimize urls and some views implement logic
