+++ articles/apps.py
@@ -4,4 +4,4 @@ from django.apps import AppConfig
 class ArticlesConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "articles"
-    verbose_name = "评论和文章"
+    verbose_name = "文章"
