+++ articles/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from .models import Article
+
+
+class ArticleSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Article
+        fields = "__all__"
