+++ articles/urls.py
@@ -1,17 +1,15 @@
 from django.urls import path
 
 from . import views
+from .views import ArticleAPIView, ArticlePageView
 
 urlpatterns = [
     # 页面路由
-    path("article", views.article_page, name="single article page"),
-    path("articles", views.articles_page, name="articles list page"),
-    path("publish", views.publish_page, name="publish page"),
+    path("", ArticlePageView.as_view(), name="article(s) page"),
     # API路由
-    path("api/articles", views.get_articles, name="get articles api"),
-    path("api/publish-article", views.publish_article, name="publish article api"),
-    path("api/article/thumb", views.thumb_article, name="thumb article api"),
-    path("api/article/viewed", views.viewed_article, name="viewed article api"),
-    path("api/comments", views.get_comments_data, name="get comments api"),
-    path("api/publish-comment", views.publish_comment, name="publish comment api"),
+    # TODO: put[更新/修改文章,当前版本未实现], delete[删除文章,当前版本未实现]
+    # 这个路由是一个标准的RESTful API，用于文章的各种操作(get[封装了单文章和文章列表获取], post[发表文章])
+    path("api", ArticleAPIView.as_view(), name="get articles api"),
+    path("api/thumb", views.thumb_article, name="thumb article api"),
+    path("api/viewed", views.viewed_article, name="viewed article api"),
 ]
