+++ comments/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class CommentsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "comments"
+    verbose_name = "评论"
